Changed doRender's return type to int16.

+++ graphics/GuiBgmapAnimatedSprite/source/GuiBgmapAnimatedSprite.c
@@ -32,13 +32,13 @@ void GuiBgmapAnimatedSprite::destructor()
 	Base::destructor();
 }
 
-uint16 GuiBgmapAnimatedSprite::doRender(int16 index, bool evenFrame)
+int16 GuiBgmapAnimatedSprite::doRender(int16 index, bool evenFrame)
 {
 	NM_ASSERT(!isDeleted(this->texture), "GuiBgmapAnimatedSprite::doRender: null texture");
 
-	uint16 result = Base::doRender(this, index, evenFrame);
+	int16 result = Base::doRender(this, index, evenFrame);
 
-	if((uint16)__NO_RENDER_INDEX == result)
+	if(__NO_RENDER_INDEX == result)
 	{
 		return __NO_RENDER_INDEX;
 	}
